requirements.txt
mysql_connector_repackaged==0.3.1
pwnedpasswords==2.0.0
python_nmap==0.7.1
Requests==2.28.2
streamlit==1.21.0
